pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eqxfer"
version = "0.1.0"
description = "Cross-tectonic earthquake magnitude transfer via split universal-physics / regional-site architecture"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "EQPaper authors" }]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.1",
    "h5py>=3.10",
    "torch>=2.2",
    "scikit-learn>=1.4",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "obspy>=1.4",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "black>=24.3",
    "mypy>=1.9",
]
analysis = [
    "reverse_geocoder>=1.5.1",
]
tracking = [
    "wandb>=0.17",
    "mlflow>=2.12",
]
viz = [
    "matplotlib>=3.8",
    "umap-learn>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["eqxfer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
